Add missing references to language-specific docs
These references are missing and the getting started guides are missing
the link to install Pack CLI. It was not obvious without the link where
to find Pack CLI. It looks like hugo doesn't actually pick up the links
in footer through the parent page, when the language-specific version of
these instructions gets broken out into separate docs.

Testing this change locally it appears to resolve the broken link.

Signed-off-by: Kingdon Barrett <kingdon@weave.works>